% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness of fit based on left-truncated regression model}
\usage{
gof(x, B = 200, Q = 1)
}
\arguments{
\item{x}{an object of class \code{trSurvfit} returned by the
\code{trSurvfit()} or the \code{trReg()} function.}

\item{B}{a integer value specifies the bootstrap size.}

\item{Q}{a integer value specifies number of breakpoints to test the linearity of the transformation model. Default value is 1.}
}
\value{
A list containing the following elements
\describe{
  \item{coefficients}{the regression coefficients of the left-truncated regression model.}
  \item{pval}{the p-value for testing if all the null hypothesis that all regression coefficients are zero.}
}
}
\description{
Provide goodness-of-fit diagnostics for the transformation model.
}
\details{
The googness of fit assessment of the transformation model focus on the structure of the
transformation model, which has the form:
\deqn{h(U) = (1 + a)^{-1} \times (h(T) + ah(X)),}
where \eqn{T} is the truncation time, \eqn{X} is the observed failure time,
\eqn{U} is the transformed truncation time that is quasi-independent from \eqn{X} and
\eqn{h(\cdot)} is a monotonic transformation function.
With the condition, \eqn{T < X}, assumed to be satisfied,
the structure of the transformation model implies
\deqn{X - T = -(1 + a) E(U) + (1 + a) X - (1 + a) \times [U - E(U)] := \beta_0 + \beta_1X + \epsilon.}
}
\examples{
data(channing, package = "boot")
chan <- subset(channing, entry < exit)
}
