% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{Surv}
\alias{Surv}
\alias{is.Surv}
\title{This is the \code{Surv} function imported from \code{survival}}
\value{
a \code{Surv} object representing left-truncated and right-censored data, as defined in the `survival` package.
}
\description{
This function is imported from the \code{survival} package,
used to create survival objects to be fitted with \code{trReg()}
and \code{trSurvfit}. See \code{\link[survival]{Surv}} for the
documentation on \code{Surv}.
}
