% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trSurv_pkg.R
\docType{package}
\name{tranSurv-package}
\alias{tranSurv}
\alias{tranSurv-package}
\title{tranSurv:Transformation Model Based Survival Curve Estimation with Dependent Left Truncation}
\description{
A package that estimates survival curve under a dependent truncation and independent right censoring via a structural transformation method.
The package also includes hypothesis test of quasi-independence based on the conditional Kendall's tau of Martin and Betensky (2005) and
two versions of the inverse probability weighted Kendall's tau of Austin and Betensky (2014).
}
\references{
Martin E. and Betensky R. A. (2005), Testing quasi-independence of failure and truncation times via conditional Kendall's tau,
\emph{Journal of the American Statistical Association}, \bold{100} (470): 484-492.

Austin, M. D. and Betensky R. A. (2014), Eliminating bias due to censoring in Kendall's tau estimators for quasi-independence of truncation and failure,
\emph{Computational Statistics & Data Analysis}, \bold{73}: 16-26.

Chiou, S., Austin, M., Qian, J. and Betensky R. A. (2016), Transformation model estimation of survival under dependent truncation and independent censoring,
\emph{Statistical Methods in Medical Research}, \bold{28} (12): 3785-3798.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stc04003/tranSurv}
  \item Report bugs at \url{https://github.com/stc04003/tranSurv/issues}
}

}
\author{
\strong{Maintainer}: Sy Han (Steven) Chiou \email{schiou@utdallas.edu}

Authors:
\itemize{
  \item Jing Qian \email{qian@schoolph.umass.edu}
}

}
