% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trReg.R
\name{trReg}
\alias{trReg}
\title{Fitting regression model via structural transformation model}
\usage{
trReg(formula, data, subset, tFun = "linear", method = c("kendall",
  "adjust"), bootSE = FALSE, control = list())
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is assumed to be a \code{survival::Surv} object with both left truncation and right censoring.
See \code{?survival::Surv} for more details.}

\item{data}{an optional data.frame in which to interpret the variables occurring
in the \code{formula}.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{tFun}{a character string specifying the transformation function or a user specified function indicating the relationship between X, T, and a.
When \code{tFun} is a character, the following are permitted:
\describe{
  \item{linear}{linear transformation structure}
  \item{log}{log-linear transformation structure}
  \item{exp}{exponential transformation structure}
}}

\item{method}{a character string specifying the underlying model. See \bold{Details}.}

\item{bootSE}{a logical value indicating whether to find the bootstrap standard error.}
}
\description{
\code{trReg} fits transformation model under dependent truncation and independent censoring via a structural transformation model.
}
