% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trSurvfit.R
\name{trSurv.control}
\alias{trSurv.control}
\title{Auxiliary for Controlling trSurvfit Fitting}
\usage{
trSurv.control(
  interval = c(-1, 20),
  lower = min(interval),
  upper = max(interval)
)
}
\arguments{
\item{interval}{a vector containing the end-points of the interval to be searched the transformation parameter.}

\item{lower}{the lower end-point of the interval to be searched.}

\item{upper}{the upper end-point of the interval to be searched.}
}
\description{
Auxiliary function as user interface for \code{trSurvfit} fitting.
}
\seealso{
\code{\link{trSurvfit}}
}
