% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness of fit based on left-truncated regression model}
\usage{
gof(x, B = 200, P = 1, Q = 0)
}
\arguments{
\item{x}{an object of class \code{trSurvfit} returned by the
\code{trSurvfit()} or the \code{trReg()} function.}

\item{B}{a integer value specifies the bootstrap size.}

\item{P}{a integer value specifies number of breakpoints to test the linearity of the transformation model. Default value is 1.}
}
\value{
A list containing the following elements
\describe{
  \item{coefficients}{the regression coefficients of the left-truncated regression model.}
  \item{pval}{the p-value for the equality of the piecewise linearity terms in the expanded model. See \bold{Details}.}
}
}
\description{
Provide goodness-of-fit diagnostics for the transformation model.
}
\details{
The googness of fit assessment of the transformation model focus on the structure of the
transformation model, which has the form:
\deqn{h(U) = (1 + a)^{-1} \times (h(T) + ah(X)),}
where \eqn{T} is the truncation time, \eqn{X} is the observed failure time,
\eqn{U} is the transformed truncation time that is quasi-independent from \eqn{X} and
\eqn{h(\cdot)} is a monotonic transformation function.
With the condition, \eqn{T < X}, assumed to be satisfied,
the structure of the transformation model implies
\deqn{X - T = -(1 + a) E(U) + (1 + a) X - (1 + a) \times [U - E(U)] := \beta_0 + \beta_1X + \epsilon.}
The regression estimates can be obtained by the left-truncated regression model (Karlsson and Lindmark, 2014).
To evaluate the goodness of fit of the transformation model,
the \code{gof()} function directly test the inearity in \eqn{X} by considering larger model that are nonlinear in \eqn{X}.
In particular, we expand the covariates \eqn{X} to \code{Q} piecewise linearity terms and test for equality of the associated coefficients.
}
\examples{
data(channing, package = "boot")
chan <- subset(channing, entry < exit)
fit <- trReg(Surv(entry, exit, cens) ~ sex, data = chan)
gof(fit, B = 25)
}
\references{
Karlsson, M., Lindmark, A. (2014) truncSP: An R Package for Estimation of Semi-Parametric Truncated Linear Regression Models, \emph{Journal of Statistical Software}, \bold{57} (14), pp 1--19.
}
