% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wKendall.R
\name{wKendall}
\alias{wKendall}
\title{Weighted conditional Kendall's tau}
\usage{
wKendall(trun, obs, delta = NULL, weights = NULL)
}
\arguments{
\item{trun}{left truncation time satisfying \code{trun} <= \code{obs}.}

\item{obs}{observed failure time, must be the same length as \code{trun}, might be right-censored.}

\item{delta}{an optional 0-1 vector of censoring indicator (0 = censored, 1 = event) for \code{obs}.
If this vector is not specified, \code{condKendall} assumes no censoring and all observed failure time
denote events.}

\item{weights}{perturbation weights.}
}
\description{
This is a temporary function, could be merged into the main function \code{condKendall}.
Give conditional Kendall's tau with the ability to take on perturbation weights.
}
